requirements.txt
arrow==0.15.2
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.3.0
binaryornot==0.4.4
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
coloredlogs==10.0
cookiecutter==1.6.0
coverage==4.5.4
docutils==0.15.2
filelock==3.0.12
future==0.18.1
hashlib-additional==1.0.1
humanfriendly==4.18
idna==2.8
importlib-metadata==0.23
isort==4.3.21
Jinja2==2.10.3
jinja2-time==0.2.0
lazy-object-proxy==1.4.1
listdir==0.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
nose==1.3.7
packaging==19.2
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.13.0
poyo==0.5.0
py==1.8.0
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.2
pyreadline==2.1
pytest==5.2.1
pytest-cov==2.8.1
python-dateutil==2.8.0
PyYAML==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
toml==0.10.0
tox==3.14.0
tqdm==4.33.0
twine==1.13.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
whichcraft==0.6.1
wrapt==1.11.2
zipfile37==0.1.3
zipp==0.6.0
